requirements.txt
discord.py>=2.3.2
google-generativeai>=0.3.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0 